+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Better encapsulated widgets selecting lyric aligner(s) and lyric fetcher(s)
+- Re-factored NUSAutoLyrixAlignOffline Aligner to properly copy temporary files and manage working directory
 - Notes are now added to processing table hinting at issues/progress with processed files.
 - Paths to individual audio files, as well as folders, is now supported.
 - Fixed bug where LyricManager would occasionally only recognize .mp3 files, and not other valid types of audio files.